[util] - separate / share-able 3d-tiles -> three conversion utils
Maybe? So far I have needed to duplicate the majority of this functionality in order to work with my custom extension? Is it too much to export the utility functions from the package as a way to keep internal/cached properties from being relied upon by external package users / exposed via types?